WIP: Implement IPlayable in SoundSprite
This is a larger update that attempts to bring the new IPlayable interface to
the SoundSprite class. This is a work in progress, and currently does not work
as intended.

Important changes:
- Abstraction of ISoundSpriteSheetTexture data type
- The texture field now has type ISoundSpriteSheetTexture
- The setTexture method has been implemented
- Formatting
- Creation of IPlaying list
- Move away from keeping node info as SoundSprite instance fields
  (WIP: still hasn't updated a bunch of "this."-stuff)
- Node creation added on a per-texture basis
- Added comments
  - Many comments are TODO-notes for future self

Also note that there are problems with the automatic file loading code as per
the current setup - I noticed this when removing the call to sound.stop() in
index.ts, which messed up my testing for the longest time because I hadn't fixed
the stop code yet.