Refactor and fix compile errors
This is a larger update, fixing numerous compile time errors. As a result of the
SoundSprite file being removed, this includes changing a lot of references to
that file and corresponding classes. In an attempt to get compilation to happen,
I made some radical changes to parts of the API dealing with creation of sounds
and whatnot. These were made in a hurry, and probably need to be revised. There
are surely problems.